話し合い画面までを作成
