+++ photo-werewolf-iOS/Data/Entity/UserEntity.swift
@@ -21,6 +21,7 @@ struct GameUser: Codable, Identifiable {
 	var userId: String
 	var name: String
 	var photoUrl: String?
+	var exchangePhotoUrl: String?
 	var role: StandardGameRole?
 }
